со светлой темой. с проведенным манки тестом. с редактированием статей расходов: не оттестировано, и фрагменты создаются с каждым действием

+++ app/src/main/java/com/jeezsoft/costcontrol/onSomeEventListener.java
@@ -1,6 +1,6 @@
 package com.jeezsoft.costcontrol;
 
 public interface onSomeEventListener {
-    public void someEvent(float sum);
+    public void someEvent(Double sum);
     public void showKeyboard(boolean show);
 }
